Initial GPU calculation implementation
A lot to be done, not yet tested, first compilable version

+++ src/data_table.hpp
@@ -11,27 +11,27 @@ public:
     // Constructors
     DataTable_t();
     ~DataTable_t();
-    DataTable_t(std::string colname, std::vector<double> x);
-    DataTable_t(std::string colname1, std::vector<double> x,
-                std::string colname2, std::vector<double> y);
+    DataTable_t(std::string colname, std::vector<float> x);
+    DataTable_t(std::string colname1, std::vector<float> x,
+                std::string colname2, std::vector<float> y);
     DataTable_t(size_t n, size_t m);
 
     // Modifiers
-    void append(std::string colname, std::vector<double> x);
+    void append(std::string colname, std::vector<float> x);
     void resize(Dim_t dim, size_t sz);
     // Acess
-    std::vector<double>& at(size_t i);
-    double               at(size_t i, size_t j);
-    double               at(std::string colname, size_t j);
+    std::vector<float>& at(size_t i);
+    float               at(size_t i, size_t j);
+    float               at(std::string colname, size_t j);
     // Queries
     size_t getColNumber(std::string colname);
     size_t getNumberRows();
     size_t getNumberCols();
-    double getVal(std::string col_out, std::string col_in, double x);
+    float getVal(std::string col_out, std::string col_in, float x);
 
     // Data containers
     std::vector<std::string> cols;
-    std::vector< std::vector<double> > data;
+    std::vector< std::vector<float> > data;
 };
 
 #endif
